% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artset1987.R
\docType{data}
\name{artset1987}
\alias{artset1987}
\title{Artset1987 data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:} {survival time in days}
}
}
\usage{
data(artset1987)
}
\description{
time
}
\examples{


data(artset1987)
head(artset1987)

}
\references{
Aarset, M. V. (1987). How to Identify a Bathtub Hazard Rate. IEEE Transactions on Reliability, 36, 106–108.
}
\keyword{dataSets}
