% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\docType{data}
\name{cycles}
\alias{cycles}
\title{cycles data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:} {survival time in days}
}
}
\usage{
data(cycles)
}
\description{
time
}
\examples{


data(cycles)
head(cycles)

}
\references{
Loprinzi CL. Laurie JA. Wieand HS. Krook JE. Novotny PJ.Kugler JW. Bartel J. Law M. Bateman M. Klatt NE.
et al.Prospective evaluation of prognostic variables from patient-completed questionnaires.
North Central Cancer Treatment Group.Journal of Clinical Oncology. 12(3):601-7, 1994.
}
\keyword{dataSets}
