% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepatitis.R
\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{Hepatitis data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:}{survival time in weeks}
\item {censured:} {censored = 0, dead = 1}
\item {group:} {control = 1, steroid = 2}
}
}
\usage{
data(hepatitis)
}
\description{

}
\examples{


data(hepatitis)
head(hepatitis)

}
\references{
Colosimo, E.A . Análise de Sobrevivência Aplicada
}
\keyword{dataSets}
