% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXI.R
\name{MaxGTDL}
\alias{MaxGTDL}
\alias{likeGTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
likeGTDL(param)

MaxGTDL(start, ...)
}
\description{
This is the main interface of the maxLik package and the function that performs the maximum probability estimate of the GTDL package
}
\examples{

## times data (from Arset, 1987))

data(artset1987)
t<-artset1987
max<-MaxGTDL(c(1,-0.05,-1))

}
\references{
How to identify bathtub hazard rate (Aarset,1987 )
Mack
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
