% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXI2.R
\name{MaxGTDL2}
\alias{MaxGTDL2}
\alias{likeGTDL2}
\title{}
\usage{
likeGTDL2(t1, censura, para)

MaxGTDL2(start, t, censur, ...)
}
\description{

}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}

Colosimo, E. A.,() Análise de sobrevivência aplicada
}
