% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXI.R
\name{MaxGTDL}
\alias{MaxGTDL}
\alias{mle.GTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
mle.GTDL(start, t, method = "BFGS")
}
\arguments{
\item{start}{vector of parameters to obtaind maximum likelihood.}

\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}
}
\description{
The maximum likelihood estimation of the GTDL distribution
}
\examples{
# times data (from Aarset, 1987))
data(artset1987)
mod <- mle.GTDL(c(1,-0.05,-1),t = artset1987)

}
\references{
Aarset, M. V. (1987). How to Identify a Bathtub Hazard Rate. IEEE Transactions on Reliability, 36, 106–108.
Mackenzie, G. (1996) Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family. Journal of the Royal Statistical Society.
Series D (The Statistician). (45). 21-34.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
