% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiGTDL.R
\name{quantile.GTDL}
\alias{quantile.GTDL}
\title{Residual value of the GTDL distribution}
\usage{
\method{quantile}{GTDL}(t, formula, param, censura)
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{param}{These are the lambda, alpha and gamma parameters of the GTDL distribution.}

\item{censura}{absence of the occurrence of the event at the time of analysis.}
}
\description{
Residual value of the GTDL distribution
}
\examples{

data(lung)
lung <- lung[-14,]
lung$ph.ecog[lung$ph.ecog==3]<-2
formula <- ~lung$sex+factor(lung$ph.ecog)+lung$age
censur <- ifelse(lung$status==1,0,1)
start <- c(0.03,0.05,-1,0.7,2,-0.1)
fit.model <- mle2.GTDL(t = lung$time,start = start,
           formula = formula,
           censur = censur)
r <- quantile.GTDL(t = lung$time,formula = formula ,param = fit.model$Coefficients[,1],
             censura = censur)
envelope.GTDL(x)
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
