% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiGTDL.R
\name{resi.GTDL}
\alias{resi.GTDL}
\alias{quantileGTDL}
\alias{envelope.GTDL}
\title{Residual value of the GTDL distribution}
\usage{
quantileGTDL(t, formula, param, censura)

envelope.GTDL(x)
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{param}{These are the lambda, alpha and gamma parameters of the GTDL distribution.}

\item{censura}{absence of the occurrence of the event at the time of analysis.}
}
\description{
Residual value of the GTDL distribution
}
\examples{

data("hepatitis")
x <- quantile.GTDL(t = hepatitis$t,param = c(1,-0.05,-1),
             censura = hepatistis$censured)
envole.GTDL(x)
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
