% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiGTDL.R
\name{q.GTDL}
\alias{q.GTDL}
\title{Residual value of the GTDL distribution}
\usage{
q.GTDL(t, formula, param, censura)
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{formula}{The structure matrix of covariates of dimension n x p}

\item{param}{These are the lambda, alpha and gamma parameters of the GTDL distribution.}

\item{censura}{absence of the occurrence of the event at the time of analysis.}
}
\description{
Residual value of the GTDL distribution
}
\examples{

data(lung)
lung <- lung[-14,]
lung$ph.ecog[lung$ph.ecog==3]<-2
t1 <- lung$time
formula1 <- ~lung$sex+factor(lung$ph.ecog)+lung$age
censur1 <- ifelse(lung$status==1,0,1)
start1 <- c(0.03,0.05,-1,0.7,2,-0.1)
fit.model1 <- mle2.GTDL(t = t1,start = start1,
           formula = formula1,
           censur = censur1)
r1 <- q.GTDL(t = t1,formula = formula1 ,param = fit.model1$Coefficients[,1],
             censura = censur1)
r1


data(tumor)
t2 <- tumor$t
formula2 <- ~tumor$group
censur2 <- tumor$censured
start2 <- c(1,-0.05,1.7)
fit.model2 <- mle2.GTDL(t = t2,start = start2,
                       formula = formula2,
                       censur = censur2)
r2 <- q.GTDL(t = t2,formula = formula2, param = fit.model2$Coefficients[,1],
            censur = censur2)
r2
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
