% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\docType{data}
\name{cycles}
\alias{cycles}
\title{cycles data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:} {Cycles to failure {in units of 10^3 cycles) of springs}
\item {censured:} {censored = 0, failure = 1}
\item {group:} {Stress {N/mm^2) of springs}
}
}
\usage{
data(cycles)
}
\description{
time
}
\examples{


data(cycles)
head(cycles)

}
\references{
Dr. Cox and D. Oakes. Analysis of Survival Data
}
\keyword{dataSets}
