% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGTDL.R
\name{fGTDL}
\alias{fGTDL}
\alias{dGTDL}
\alias{hGTDL}
\alias{sGTDL}
\alias{rGTDL}
\alias{fires}
\title{The GTDL distribution}
\source{
[d-p-q-r]GTDL are calculated directly from the definitions.
}
\usage{
dGTDL(t, param, log = FALSE)

hGTDL(t, param)

sGTDL(t, param)

rGTDL(n, param)
}
\arguments{
\item{t}{vector of integer positive quantile.}

\item{param}{parameters (alpha and gamma are scalars, lambda non-negative).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations.}
}
\value{
\code{dGTDL} gives the density function, \code{hGTDL} gives the failure function, \code{sGTDL} gives the survival function and \code{rGTDL} generates random samples.

Invalid arguments will return an error message.
}
\description{
Density function, survival function, failure function and random 
generation for the GTDL distribution.
}
\details{
\itemize{
\item Density function
\deqn{f(t\mid \boldsymbol{\theta})=\lambda\left(\frac{\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}{1+\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}\right)\times\left(\frac{1+\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}{1+\exp\{\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}\right)^{-\lambda/\alpha}}

\item Survival function
\deqn{S(t \mid \boldsymbol{\theta})=\left(\frac{1+\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}{1+\exp\{\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}\right)^{-\lambda/\alpha}}

\item Failure function
\deqn{h(t\mid\boldsymbol{\theta})=\lambda\left(\frac{\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}{1+\exp\{\alpha{t}+\boldsymbol{X}^{\top}\boldsymbol{\beta}\}}\right)}

}
}
\examples{

### Example 1

library(GTDL)
library(ggplot2)
library(tidyverse)

t <- seq(0, 20, by = 0.01)
theta <- c(1.00, -0.05, -1.00)
y1 <- hGTDL(t, theta)
y2 <- sGTDL(t, theta)
y3 <- dGTDL(t, theta)

df <- data.frame(t, y1, y2, y3)

df \%>\% ggplot() + geom_line(aes(t,y = y1, col = 'failure function'), linetype = 1) +
geom_line(aes(t, y = y2, col = 'survival function'), linetype = 2 ) +
geom_line(aes(t, y = y3, col = 'density function'), linetype = 3)+
theme(legend.position = "right", plot.title = element_text(size = rel(1.2), lineheight = .9, family = "Calibri", face = "bold", colour = "brown"))


### Example 2

library(GTDL)
library(ggplot2)
library(tidyverse)

t <- seq(0, 20, by = 0.01)
theta1 <- c(1.00, 0.50, -1.00)
theta2 <- c(1.00, -0.50, 1.00)
theta3 <- c(1.00, -0.25, 1.00)
theta4 <- c(1.00, -0.06, -1.60)
theta5 <- c(1.00, 0.25, -1.00)

y1 <- hGTDL(t, theta1)
y2 <- hGTDL(t, theta2)
y3 <- hGTDL(t, theta3)
y4 <- hGTDL(t, theta4)
y5 <- hGTDL(t, theta5)

df <- data.frame(t, y1, y2, y3, y4, y5)

df \%>\% ggplot() + geom_line(aes(t, y = y1), linetype = 2) + 
 geom_line(aes(t, y = y2), linetype = 3) + geom_line(aes(t, y = y3), linetype = 4) + 
 geom_line(aes(t, y = y4), linetype = 5) + geom_line(aes(t, y = y5), linetype = 6) + 
 labs(title = "", y = "")
}
\references{
\itemize{
\item Mackenzie, G. (1996). Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family. Journal of the Royal Statistical Society. 
Series D (The Statistician). 45. 21-34. 
}
}
