% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2.R
\name{mle2.GTDL}
\alias{mle2.GTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
mle2.GTDL(t, start, formula, censur, method = "BFGS")
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{start}{vector of parameters to obtaind maximum likelihood.}

\item{formula}{The structure matrix of covariates of dimension n x p}

\item{censur}{non-negative random variable that represents whether the sample is censored or not.}

\item{method}{The method to be used}
}
\description{
Maximum probability estimate of the GTDL package
}
\examples{

### Example 1

require(survival)
data(lung)

lung <- lung[-14,]
lung$ph.ecog[lung$ph.ecog==3]<-2
t1 <- lung$time
start1 <- c(0.03,0.05,-1,0.7,2,-0.1)
formula1 <- ~lung$sex+factor(lung$ph.ecog)+lung$age
censur1 <- ifelse(lung$status==1,0,1)
fit.model1 <- mle2.GTDL(t = t1,start = start1,
                     formula = formula1,
                     censur = censur1)
fit.model1

### Example 2

data(tumor)
t2 <- tumor$time
start2 <- c(1,-0.05,1.7)
formula2 <- ~tumor$group
censur2 <- tumor$censured
fit.model2 <- mle2.GTDL(t = t2,start = start2,
                       formula = formula2,
                       censur = censur2)
fit.model2

}
\references{
Colosimo, E. A and  Giolo, S. R. Análise de sobrevivência aplicada.  Edgard Blucher: São Paulo. 2006.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
