% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTDL.R
\name{GTDL}
\alias{GTDL}
\alias{dGTDL}
\title{The Distribution GTDL}
\usage{
dGTDL(t, lambda, alpha, gamma, log = FALSE)
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{lambda}{non-negative.}

\item{alpha, gamma}{scalars.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Density, survival function, failure function and random generation for the GTDL distribution.
}
\examples{
t <- seq(0,20,by = 0.01)
lambda <- 1
alpha <- -0.05
gamma <- -1
denGTDL<- dGTDL(t,lambda,alpha,gamma,log = FALSE)
plot(x = t,y = denGTDL)
}
\references{
Mackenzie,G. ,(2016),Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
