% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTDL.R
\name{GTDL}
\alias{GTDL}
\alias{dGTDL}
\alias{hGTDL}
\alias{sGTDL}
\alias{rGTDL}
\title{The Distribution GTDL}
\usage{
dGTDL(t, param, log = FALSE)

hGTDL(t, param)

sGTDL(t, param)

rGTDL(n, param)
}
\arguments{
\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{alpha, gamma}{scalars.}

\item{lambda}{non-negative.}

\item{pcensura}{sample censorship rate}
}
\description{
Density, survival function, failure function and random generation for the GTDL distribution.
}
\examples{

library(GTDL)
t <- seq(0,20,by = 0.1)
lambda <- 1.00
alpha <- -0.05
gamma <- -1.00
y1 <- hGTDL(t,lambda,alpha,gamma)
y2 <- sGTDL(t,lambda,alpha,gamma)
y3 <- dGTDL(t,lambda,alpha,gamma,log = FALSE)
tt <- as.matrix(cbind(t,t,t))
yy <- as.matrix(cbind(y1,y2,y3))
matplot(tt,yy,type="l",xlab="time",ylab="",lty = 1:3,col=1:3,lwd=2)


y1 <- hGTDL(t,1,0.5,-1.0)
y2 <- hGTDL(t,1,0.25,-1.0)
y3 <- hGTDL(t,1,-0.25,1.0)
y4 <- hGTDL(t,1,-0.50,1.0)
y5 <- hGTDL(t,1,-0.06,-1.6)
tt <- as.matrix(cbind(t,t,t,t,t))
yy <- as.matrix(cbind(y1,y2,y3,y4,y5))
matplot(tt,yy,type="l",xlab="time",ylab="Hazard function",lty = 1:3,col=1:3,lwd=2)

}
\references{
Mackenzie,G.,(1996).Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family. Journal of the Royal Statistical Society. Series D (The Statistician), Vol. 45, No. 1, pp. 21-34.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
