% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepatitis.R
\docType{data}
\name{hepatitis}
\alias{hepatitis}
\title{Hepatitis data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:}{survival time in weeks}
\item {censured:} {censored = 0, dead = 1}
\item {group:} {control = 0, steroid = 1}
}
}
\usage{
data(hepatitis)
}
\description{
A randomized clinical trial was carried out to investigate the effect of therapy with steroid in the treatment of acute
viral hepatitis ( Gregory et al., 1975). Twenty and nine patients with this disease were randomized to receive a placebo or the steroid
treatment. Each patient was followed up for 16 weeks or until death( event of interest) or even loss of follow-up.
}
\examples{


data(hepatitis)
head(hepatitis)

}
\references{
Colosimo, E.A . Análise de Sobrevivência Aplicada
}
\keyword{dataSets}
