% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.R
\name{max.GTDL}
\alias{max.GTDL}
\alias{like2}
\title{Maximum probability estimate of the GTDL package}
\usage{
like2(t, censur, para)

\method{max}{GTDL}(start, t, censur)
}
\description{
Maximum probability estimate of the GTDL package
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}

Colosimo, E. A.,() Análise de sobrevivência aplicada
}
