% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.R
\name{max.GTDL}
\alias{max.GTDL}
\alias{maxGTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
maxGTDL(start, t, censur, method = "BFGS")
}
\arguments{
\item{start}{vector of parameters to obtaind maximum likelihood.}

\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}

\item{censur}{non-negative random variable that represents whether the sample is censored or not.}
}
\description{
Maximum probability estimate of the GTDL package
}
\examples{

data(hepatis)
maxparam <- max.GTDL(start = c(1,-0.05,-1),t = hepatitis$t, censur = hepatitis$censured)
maxparam
}
\references{
Colosimo, E. A and  Giolo, S. R. Análise de sobrevivência aplicada.  Edgard Blucher: São Paulo. 2006.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
