% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXI.R
\name{MaxGTDL}
\alias{MaxGTDL}
\alias{likeGTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
likeGTDL(param, t1)

MaxGTDL(start, t, ...)
}
\description{
This is the main interface of the maxLik package and the function that performs the maximum probability estimate of the GTDL package
}
\examples{

## times data (from Arset, 1987))

data(artset1987)
t<-artset1987
max<-MaxGTDL(c(1,-0.05,-1))

}
\references{
Aarset,M. V.,(1987). How to identify bathtub hazard rate. IEEE transactions on reliability.
Mackenzie,G.,(1996). Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family. Journal of the Royal Statistical Society. Series D (The Statistician), (45), 21-34.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
