% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXI.R
\name{MaxGTDL}
\alias{MaxGTDL}
\title{Maximum probability estimate of the GTDL package}
\usage{
MaxGTDL(start, t, ...)
}
\arguments{
\item{start}{vector of parameters to obtaind maximum likelihood.}

\item{t}{non-negative random variable representing the failure time and leave the snapshot failure rate, or danger.}
}
\description{
This is the main interface of the maxLik package and the function that performs the maximum probability estimate of the GTDL package
}
\examples{

# times data (from Aarset, 1987))

data(artset1987)
mod <- MaxGTDL(c(1,-0.05,-1))

}
\references{
Aarset,M. V. How to identify bathtub hazard rate. IEEE transactions on reliability. Two full days, Washington, DC area, v. r-36, n. 1 1987.
Mackenzie, G. Regression Models for Survival Data: The Generalized Time-Dependent Logistic Family. Journal of the Royal Statistical Society. Series D (The Statistician),  v. 45, n. 1, p. 21-34. 1996.
}
\author{
Jalmar M. F. Carrasco \email{carrascojalmar@gmail.com}

Luciano S. Santos \email{lucianno0800@gmail.com}
}
