% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor.R
\docType{data}
\name{tumor}
\alias{tumor}
\title{Tumor data}
\format{
This data frame contains the following columns:
\itemize{
\item {t:} {survival time in days}
\item {censured:} {censored = 0, dead = 1}
\item {group:} {large tumor = 0, small tumor = 1}
}
}
\usage{
data(tumor)
}
\description{
Times (in days) of patients in ovarian cancer study
}
\examples{


data(tumor)
head(tumor)

}
\references{
Colosimo, E. A and  Giolo, S. R. Análise de sobrevivência aplicada.  Edgard Blucher: São Paulo. 2006.
}
\keyword{dataSets}
